requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
python-dotenv==1.0.0

# LLM and ML
transformers==4.35.0
torch==2.1.0
accelerate==0.23.0
bitsandbytes==0.41.1
sentence-transformers==2.2.2
scikit-learn==1.3.2
xgboost==2.0.1
lightgbm==4.1.0
tensorflow==2.14.0
keras==2.14.0
statsmodels==0.14.0
langchain==0.0.312
langchain-community==0.0.10
openai==1.3.5
tiktoken==0.5.1
llama-cpp-python==0.2.11
chromadb==0.4.18

# Data Processing
pandas==2.1.1
numpy==1.26.0
pyarrow==13.0.0
polars==0.19.8
pyspark==3.5.0

# Financial Libraries
yfinance==0.2.31
alpha_vantage==2.3.1
ccxt==4.1.22
ta-lib==0.4.28
pyfolio==0.9.2
empyrical==0.5.5
risk-models==0.1.6
finrl==0.3.5
quantlib-python==1.29
finquant==0.2.3

# Streaming and Messaging
confluent-kafka==2.2.0
pika==1.3.2
redis==5.0.1
websockets==12.0

# Database
sqlalchemy==2.0.22
pymongo==4.5.0
neo4j==5.13.0
clickhouse-driver==0.2.6
influxdb-client==1.37.0

# Graph Analytics
networkx==3.2.1
igraph==0.10.8

# API Clients
requests==2.31.0
aiohttp==3.8.6
httpx==0.25.0

# Authentication & Security
pyjwt==2.8.0
passlib==1.7.4
cryptography==41.0.4
bcrypt==4.0.1

# Blockchain
web3==6.11.1
eth-account==0.9.0
hyperledger-fabric-sdk==0.8.1

# Monitoring and Logging
opentelemetry-api==1.20.0
opentelemetry-sdk==1.20.0
prometheus-client==0.17.1
elasticsearch==8.10.1
loguru==0.7.2

# Testing
pytest==7.4.2
pytest-asyncio==0.21.1
locust==2.18.0

# Development
black==23.10.0
isort==5.12.0
mypy==1.6.1
flake8==6.1.0

# Web UI Dependencies (for backend integration)
starlette==0.27.0
jinja2==3.1.2
itsdangerous==2.1.2
